# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Etana Open Api Documentation",
        "description": "",
        "version": ""
    },
    "servers": [
        {
            "url": "https://utajeer.com/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/userWishList": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Wish List",
                "description": "return array",
                "operationId": "userWishList",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Wish List.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Empty array",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/userCartList": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Cart",
                "description": "return array",
                "operationId": "userCartList",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Cart list.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Empty array",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/contact_us": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Contact Us",
                "description": "return string",
                "operationId": "contact_us",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Your query has been submitted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/addUserCardDetails": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "add User Card Details",
                "description": "return string",
                "operationId": "addUserCardDetails",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "card_no": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "expires_on": {
                                        "type": "string"
                                    },
                                    "cvv": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Added User Card Details Successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/editUserCardDetails": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Edit Billing Address",
                "description": "return string",
                "operationId": "editUserCardDetails",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "card_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "card_no": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "expires_on": {
                                        "type": "string"
                                    },
                                    "cvv": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated Card Details Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/deleteUserCardDetails": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Delete User Card Details",
                "description": "return string",
                "operationId": "deleteUserCardDetails",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "card_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deleted Card Details Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/listUserCardDetails": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "List of Card Detail",
                "description": "return array",
                "operationId": "listUserCardDetails",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of Card Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Empty array",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}